apps/converter/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pdf2docx==0.5.8
boto3==1.34.131
python-magic==0.4.27
pydantic==2.7.4
pydantic-settings==2.3.4
httpx==0.27.0
